Added MultiThreadCoord3DCallable. Added jzy3d support.

+++ examples/salesmanpath/SalesmanPopulation.java
@@ -41,7 +41,7 @@ public class SalesmanPopulation extends Population {
     /**
      * Constructor for SalesmanPopulation. The id and all rates can be customized, and tournamentSize for the tournament selection.
      */
-    public SalesmanPopulation(String id, int tournamentSize, int populationSize, float mutationRate, float crossoverRate, float elitismRate) {
+    public SalesmanPopulation(String id, int tournamentSize, int populationSize, double mutationRate, double crossoverRate, double elitismRate) {
         super(id, populationSize, mutationRate, crossoverRate, elitismRate, generateRandomChromosome(), Population.tournamentSelectMin(tournamentSize), defaultCrossoverInterface(), defaultMutationInterface(), evolveToMax());
         //this.tournamentSize = tournamentSize;
     }
@@ -50,9 +50,9 @@ public class SalesmanPopulation extends Population {
             String id,
             //int tournamentSize,
             int populationSize,
-            float mutationRate,
-            float crossoverRate,
-            float elitismRate,
+            double mutationRate,
+            double crossoverRate,
+            double elitismRate,
             FunctionalChromosomeGenerator chromosomeGenerator,
             FunctionalSelectionInterface selectionInterface,
             FunctionalCrossoverInterface crossoverInterface,
